Moved file to correct place and fixed bug
