+++ src/main/resources/fxml_files/IgnoredTransactionPopup.fxml
@@ -12,7 +12,7 @@
 <?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
-<?import ledger.user_interface.ui_controllers.component.tablecolumn.IgnoredTransactionTableView?>
+<?import ledger.user_interface.ui_controllers.component.IgnoredTransactionTableView?>
 
 <fx:root fx:id="ignoredTransGridPane" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" minHeight="-Infinity" minWidth="-Infinity" prefHeight="220.0" prefWidth="558.0" style="-fx-background-color: #ffffff;" type="GridPane" xmlns="http://javafx.com/javafx/8.0.102-ea" xmlns:fx="http://javafx.com/fxml/1">
     <columnConstraints>
